+++ lib/avromatic.rb
@@ -1,4 +1,3 @@
-require 'avromatic/virtus_performance_patch'
 require 'avromatic/version'
 require 'avro_schema_registry-client'
 require 'avromatic/model'
